Some networking refactoring
