+++ game/components/actor/actor_networking.py
@@ -1,8 +1,8 @@
 from game.ecs import Component
 import game.components as C
-from ..networking.networking import Networking
+from ..networking.network_behavior import NetworkBehavior
 
-class ActorNetworking(Component, Networking):
+class ActorNetworking(Component, NetworkBehavior):
   def __init__(self):
     super().__init__()
     self.require(C.PositionSyncing, C.SpriteSyncing, C.StatsSyncing)
