+++ game/components/networking/__init__.py
@@ -1,5 +1,5 @@
+from game.components.networking.network_behavior import NetworkBehavior
 from game.components.networking.networking import Networking
-from game.components.networking.networked import Networked
 from game.components.networking.server_manager import ServerManager
 from game.components.networking.client_manager import ClientManager
 from game.components.networking.server_player import ServerPlayer
